fix: address extra field filter review feedback
