+++ spoolman/database/filament.py
@@ -12,8 +12,17 @@ from sqlalchemy.orm import contains_eager, joinedload
 
 from spoolman.api.v1.models import EventType, Filament, FilamentEvent, MultiColorDirection
 from spoolman.database import models, vendor
-from spoolman.database.utils import SortOrder
+from spoolman.database.extra_field_query import apply_extra_field_filters_and_sort
+from spoolman.database.utils import (
+    SortOrder,
+    add_where_clause_int_in,
+    add_where_clause_int_opt,
+    add_where_clause_str,
+    add_where_clause_str_opt,
+    parse_nested_field,
+)
 from spoolman.exceptions import ItemDeleteError, ItemNotFoundError
+from spoolman.extra_field_registry import EntityType
 from spoolman.math import delta_e, hex_to_rgb, rgb_to_lab
 from spoolman.ws import websocket_manager
 
@@ -107,17 +116,6 @@ async def find(
 
     Returns a tuple containing the list of items and the total count of matching items.
     """
-    # Import here to avoid circular imports
-    from spoolman.database.utils import (
-        add_where_clause_int_in,
-        add_where_clause_int_opt,
-        add_where_clause_str,
-        add_where_clause_str_opt,
-        add_where_clause_extra_field,
-        add_order_by_extra_field,
-        parse_nested_field,
-    )
-    
     stmt = (
         select(models.Filament)
         .options(contains_eager(models.Filament.vendor))
@@ -134,60 +132,31 @@ async def find(
 
     total_count = None
 
-    if limit is not None:
-        total_count_stmt = stmt.with_only_columns(func.count(), maintain_column_froms=True)
-        total_count = (await db.execute(total_count_stmt)).scalar()
-
-        stmt = stmt.offset(offset).limit(limit)
-
-    # Apply extra field filters if provided
-    if extra_field_filters:
-        # Get all extra fields for filaments
-        from spoolman.extra_fields import EntityType, ExtraFieldType, get_extra_fields
-
-        extra_fields = await get_extra_fields(db, EntityType.filament)
-        extra_fields_dict = {field.key: field for field in extra_fields}
-
-        for field_key, value in extra_field_filters.items():
-            if field_key in extra_fields_dict:
-                field = extra_fields_dict[field_key]
-                stmt = add_where_clause_extra_field(
-                    stmt,
-                    models.Filament,
-                    EntityType.filament,
-                    field_key,
-                    field.field_type,
-                    value,
-                    field.multi_choice if field.field_type == ExtraFieldType.choice else None,
-                )
+    stmt = await apply_extra_field_filters_and_sort(
+        db=db,
+        stmt=stmt,
+        base_obj=models.Filament,
+        entity_type=EntityType.filament,
+        extra_field_filters=extra_field_filters,
+        sort_by=sort_by,
+    )
 
     if sort_by is not None:
         for fieldstr, order in sort_by.items():
             # Check if this is a custom field sort
             if fieldstr.startswith("extra."):
-                field_key = fieldstr[6:]  # Remove "extra." prefix
-                
-                # Get the field definition
-                from spoolman.extra_fields import EntityType, get_extra_fields
-                
-                extra_fields = await get_extra_fields(db, EntityType.filament)
-                extra_field = next((f for f in extra_fields if f.key == field_key), None)
-                
-                if extra_field:
-                    stmt = add_order_by_extra_field(
-                        stmt,
-                        models.Filament,
-                        EntityType.filament,
-                        field_key,
-                        extra_field.field_type,
-                        order
-                    )
-            else:
-                field = parse_nested_field(models.Filament, fieldstr)
-                if order == SortOrder.ASC:
-                    stmt = stmt.order_by(field.asc())
-                elif order == SortOrder.DESC:
-                    stmt = stmt.order_by(field.desc())
+                continue
+
+            field = parse_nested_field(models.Filament, fieldstr)
+            if order == SortOrder.ASC:
+                stmt = stmt.order_by(field.asc())
+            elif order == SortOrder.DESC:
+                stmt = stmt.order_by(field.desc())
+
+    if limit is not None:
+        total_count_stmt = stmt.with_only_columns(func.count(), maintain_column_froms=True).order_by(None)
+        total_count = (await db.execute(total_count_stmt)).scalar()
+        stmt = stmt.offset(offset).limit(limit)
 
     rows = await db.execute(
         stmt,
